moved most types to specific types file to be exported
