+++ src/resources/Account/Payoffs.ts
@@ -1,17 +1,6 @@
-import Resource, { IResourceError, TResourceStatus, IResourceListOpts } from "../../resource";
+import Resource from "../../resource";
 import Configuration from "../../configuration";
-
-export interface IAccountPayoff {
-  id: string,
-  account_id: string,
-  status: TResourceStatus,
-  amount: number | null,
-  term: number | null,
-  per_diem_amount: number | null,
-  error: IResourceError | null,
-  created_at: string,
-  updated_at: string,
-};
+import type { IAccountPayoff } from "./types";
 
 export default class AccountPayoffs extends Resource {
   constructor(config: Configuration) {
