+++ src/resources/Account/index.ts
@@ -1,81 +1,21 @@
-import Resource, { IRequestConfig, IResourceListOpts, IResourceError } from '../../resource';
+import Resource, { IRequestConfig } from '../../resource';
 import Configuration from '../../configuration';
-import AccountCardBrand, { IAccountCardBrand } from './CardBrands';
-import AccountPayoffs, { IAccountPayoff } from './Payoffs';
-import AccountUpdates, { IAccountUpdate } from './Updates';
-import AccountBalances, { IAccountBalance } from './Balances';
-import AccountSensitive, { IAccountSensitive } from './Sensitive';
-import AccountTransactions, { IAccountTransaction } from './Transactions';
+import AccountCardBrand from './CardBrands';
+import AccountPayoffs from './Payoffs';
+import AccountUpdates from './Updates';
+import AccountBalances from './Balances';
+import AccountSensitive from './Sensitive';
+import AccountTransactions from './Transactions';
 import AccountSubscriptions from './Subscriptions';
-import AccountVerificationSession, { IAccountVerificationSession } from './VerificationSessions';
+import AccountVerificationSession from './VerificationSessions';
 import type {
-  IAccountACH,
-  IAccountLiability,
-  TAccountExpandableFields,
-  TAccountProducts,
-  TAccountStatuses,
-  TAccountSubscriptionTypes,
-  TAccountTypes,
-  TAccountLiabilityTypes,
-  TAccountOwnership,
+  IAccount,
+  IAccountListOpts,
+  IACHCreateOpts,
+  ILiabilityCreateOpts,
+  IAccountWithdrawConsentOpts,
+  TAccountExpandableFields
 } from './types';
-
-export interface IAccountCreateOpts {
-  holder_id: string;
-  metadata?: {};
-};
-
-export interface IACHCreateOpts extends IAccountCreateOpts {
-  ach: IAccountACH;
-};
-
-export interface ILiabilityCreateOpts extends IAccountCreateOpts {
-  liability: {
-    mch_id: string;
-    account_number?: string;
-    number?: string;
-  }
-};
-
-export interface IAccountListOpts<T extends TAccountExpandableFields> extends IResourceListOpts {
-  status?: string | null;
-  type?: string | null;
-  holder_id?: string | null;
-  expand?: T[];
-  'liability.mch_id'?: string | null;
-  'liability.type'?: TAccountLiabilityTypes | null;
-  'liability.ownership'?: TAccountOwnership | null;
-};
-
-export interface IAccountWithdrawConsentOpts {
-  type: 'withdraw';
-  reason: 'holder_withdrew_consent' | null;
-};
-
-export interface IAccount {
-  id: string;
-  holder_id: string;
-  status: TAccountStatuses;
-  type: TAccountTypes | null;
-  ach?: IAccountACH | null;
-  liability?: IAccountLiability | null;
-  products: TAccountProducts[];
-  restricted_products: TAccountProducts[];
-  subscriptions?: TAccountSubscriptionTypes[];
-  available_subscriptions?: TAccountSubscriptionTypes[];
-  restricted_subscriptions?: TAccountSubscriptionTypes[];
-  sensitive?: string | IAccountSensitive | null;
-  balance?: string | IAccountBalance | null;
-  card_brand?: string | IAccountCardBrand | null;
-  payoff?: string | IAccountPayoff | null;
-  transactions?: string | IAccountTransaction[] | null;
-  update?: string | IAccountUpdate | null;
-  latest_verification_session?: string | IAccountVerificationSession | null;
-  error: IResourceError | null;
-  created_at: string;
-  updated_at: string;
-  metadata: {} | null;
-};
 
 export class AccountSubResources {
   balances: AccountBalances;
