+++ src/resources/Account/types.ts
@@ -1,3 +1,17 @@
+import type {
+  IResourceError,
+  TResourceStatus,
+  IResourceListOpts,
+} from '../../resource';
+import type {
+  IPlaidBalance,
+  IPlaidTransaction,
+  IMXAccount,
+  IMXTransaction,
+  ITellerBalance,
+  ITellerTransaction,
+} from './externalTypes';
+
 export const AccountTypes = {
   ach: 'ach',
   liability: 'liability',
@@ -208,3 +222,372 @@ export interface IAccountACH {
   number: string;
   type: TAchAccountSubTypes;
 };
+
+export interface IAccountBalance {
+  id: string;
+  account_id: string;
+  status: TResourceStatus;
+  amount: number | null;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IAccountCardBrandInfo {
+  art_id: string;
+  url: string;
+  name: string;
+};
+
+export interface IAccountCardBrand {
+  id: string;
+  account_id: string;
+  network: string | null;
+  issuer: string | null;
+  last4: string | null;
+  brands: IAccountCardBrandInfo[];
+  status: 'completed' | 'failed';
+  shared: boolean;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IAccountPayoff {
+  id: string,
+  account_id: string,
+  status: TResourceStatus,
+  amount: number | null,
+  term: number | null,
+  per_diem_amount: number | null,
+  error: IResourceError | null,
+  created_at: string,
+  updated_at: string,
+};
+
+export const AccountSensitiveFields = {
+  auto_loan_number: 'auto_loan.number',
+  mortgage_number: 'mortgage.number',
+  personal_loan_number: 'personal_loan.number',
+  credit_card_number: 'credit_card.number',
+  credit_card_billing_zip_code: 'credit_card.billing_zip_code',
+  credit_card_exp_month: 'credit_card.exp_month',
+  credit_card_exp_year: 'credit_card.exp_year',
+  credit_card_cvv: 'credit_card.cvv',
+} as const;
+
+export type TAccountSensitiveFields = typeof AccountSensitiveFields[keyof typeof AccountSensitiveFields];
+
+export interface IAccountSensitiveLoan {
+  number: string;
+};
+
+export interface IAccountSensitiveCreditCard {
+  number: string | null;
+  billing_zip_code: string | null;
+  exp_month: string | null;
+  exp_year: string | null;
+  cvv: string | null;
+};
+
+export interface IAccountSensitive {
+  id: string;
+  account_id: string;
+  type: TAccountLiabilityTypes;
+  auto_loan?: IAccountSensitiveLoan;
+  credit_card?: IAccountSensitiveCreditCard;
+  mortgage?: IAccountSensitiveLoan;
+  personal_loan?: IAccountSensitiveLoan;
+  status: 'completed' | 'failed';
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IAccountSensitiveCreateOpts {
+  expand: TAccountSensitiveFields[];
+};
+
+export const AccountSubscriptionStatuses = {
+  active: 'active',
+  inactive: 'inactive',
+};
+
+export type TAccountSubscriptionStatuses = keyof typeof AccountSubscriptionStatuses;
+
+export interface IAccountSubscription {
+  id: string;
+  name: TAccountSubscriptionTypes;
+  status: TAccountSubscriptionStatuses;
+  latest_request_id: string | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IAccountSubscriptionsResponse {
+  transactions?: IAccountSubscription;
+  update?: IAccountSubscription;
+  'update.snapshot'?: IAccountSubscription;
+};
+
+export interface IAccountSubscriptionCreateOpts {
+  enroll: TAccountSubscriptionTypes;
+};
+
+export interface IAccountUpdate {
+  id: string;
+  status: TResourceStatus;
+  account_id: string;
+  source: TAccountUpdateSources;
+  type: TAccountLiabilityTypes;
+  auto_loan?: IAccountLiabilityAutoLoan;
+  credit_card?: IAccountLiabilityCreditCard;
+  collection?: IAccountLiabilityCollection;
+  mortgage?: IAccountLiabilityMortgage;
+  personal_loan?: IAccountLiabilityPersonalLoan;
+  student_loans?: IAccountLiabilityStudentLoans;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export const AccountCurrencyTypes = {
+  USD: 'USD',
+};
+
+export type TAccountCurrencyTypes = keyof typeof AccountCurrencyTypes;
+
+export const AccountTransactionStatuses = {
+  cleared: 'cleared',
+  auth: 'auth',
+  refund: 'refund',
+  unknown: 'unknown',
+};
+
+export type TAccountTransactionStatuses = keyof typeof AccountTransactionStatuses;
+
+export interface IAccountTransactionMerchant {
+  name: string;
+  category_code: string;
+  city: string;
+  state: string;
+  country: string;
+  acquirer_bin: string;
+  acquirer_card_acceptor_id: string;
+};
+
+export interface IAccountTransactionNetworkData {
+  visa_merchant_id: string | null;
+  visa_merchant_name: string | null;
+  visa_store_id: string | null;
+  visa_store_name: string | null;
+};
+
+export interface IAccountTransaction {
+  id: string;
+  account_id: string;
+  merchant: IAccountTransactionMerchant;
+  network: string;
+  network_data: IAccountTransactionNetworkData | null;
+  amount: number;
+  currency: TAccountCurrencyTypes;
+  billing_amount: number;
+  billing_currency: TAccountCurrencyTypes;
+  status: TAccountTransactionStatuses;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export const AccountVerificationSessionStatuses = {
+  pending: 'pending',
+  in_progress: 'in_progress',
+  verified: 'verified',
+  failed: 'failed',
+};
+
+export type TAccountVerificationSessionStatuses = keyof typeof AccountVerificationSessionStatuses;
+
+export const AccountVerificationSessionTypes = {
+  micro_deposits: 'micro_deposits',
+  plaid: 'plaid',
+  mx: 'mx',
+  teller: 'teller',
+  standard: 'standard',
+  instant: 'instant',
+  pre_auth: 'pre_auth',
+};
+
+export type TAccountVerificaionSessionTypes = keyof typeof AccountVerificationSessionTypes;
+
+export const AccountVerificationCheckStatusTypes = {
+  pass: 'pass',
+  fail: 'fail',
+};
+
+export type TAccountVerificationPassFail = keyof typeof AccountVerificationCheckStatusTypes;
+
+export interface IAccountVerificationSessionMicroDeposits {
+  amounts: number[] | [];
+};
+
+export interface IAccountVerificaitonSessionPlaid {
+  balances: IPlaidBalance | {};
+  transactions: IPlaidTransaction[] | [];
+};
+
+export interface IAccountVerificationSessionMX {
+  account: IMXAccount | {};
+  transactions: IMXTransaction[] | [];
+};
+
+export interface IAccountVerificationSessionTeller {
+  balances: ITellerBalance | {};
+  transactions: ITellerTransaction[] | [];
+};
+
+// TODO: Implement the following interfaces
+export interface IAccountVerificationSessionTrustProvisioner {};
+export interface IAccountVerificationSessionAutoVerify {};
+export interface IAccountVerificationSessionThreeDS {};
+export interface IAccountVerificationSessionIssuer {};
+
+export interface IAccountVerificationSessionStandard {
+  number: string;
+};
+
+export interface IAccountVerificationSessionInstant {
+  exp_year?: string | null;
+  exp_month?: string | null;
+  exp_check?: TAccountVerificationPassFail | null;
+  number?: string | null;
+};
+
+export interface IAccountVerificationSessionPreAuth extends IAccountVerificationSessionInstant {
+  cvv?: string | null;
+  cvv_check?: TAccountVerificationPassFail | null;
+  billing_zip_code?: string | null;
+  billing_zip_code_check?: TAccountVerificationPassFail | null;
+  pre_auth_check?: TAccountVerificationPassFail | null;
+};
+
+export interface IAccountVerificationSessionCreateOpts {
+  type: TAccountVerificaionSessionTypes;
+};
+
+export interface IAccountVerificationSessionMicroDepositsUpdateOpts {
+  micro_deposits: IAccountVerificationSessionMicroDeposits;
+};
+
+export interface IAccountVerificationSessionPlaidUpdateOpts {
+  plaid: IAccountVerificaitonSessionPlaid;
+};
+
+export interface IAccountVerificationSessionMXUpdateOpts {
+  mx: IAccountVerificationSessionMX;
+};
+
+export interface IAccountVerificationSessionTellerUpdateOpts {
+  teller: IAccountVerificationSessionTeller;
+};
+
+export interface IAccountVerificationSessionStandardUpdateOpts {
+  standard: IAccountVerificationSessionStandard;
+};
+
+export interface IAccountVerificationSessionInstantUpdateOpts {
+  instant: IAccountVerificationSessionInstant;
+};
+
+export interface IAccountVerificationSessionPreAuthUpdateOpts {
+  pre_auth: IAccountVerificationSessionPreAuth
+};
+
+export type IAccountVerificationSessionUpdateOpts =
+  | IAccountVerificationSessionMicroDepositsUpdateOpts
+  | IAccountVerificationSessionPlaidUpdateOpts
+  | IAccountVerificationSessionMXUpdateOpts
+  | IAccountVerificationSessionTellerUpdateOpts
+  | IAccountVerificationSessionStandardUpdateOpts
+  | IAccountVerificationSessionInstantUpdateOpts
+  | IAccountVerificationSessionPreAuthUpdateOpts;
+
+export interface IAccountVerificationSession {
+  id: string;
+  account_id: string;
+  status: TAccountVerificationSessionStatuses;
+  type: TAccountVerificaionSessionTypes;
+  error: IResourceError | null;
+  plaid?: IAccountVerificaitonSessionPlaid | null;
+  mx?: IAccountVerificationSessionMX | null;
+  teller?: IAccountVerificationSessionTeller | null;
+  micro_deposits?: IAccountVerificationSessionMicroDeposits | null;
+  trusted_provisioner?: IAccountVerificationSessionTrustProvisioner | null;
+  auto_verify?: IAccountVerificationSessionAutoVerify | null;
+  standard?: IAccountVerificationSessionStandard | null;
+  instant?: IAccountVerificationSessionInstant | null;
+  pre_auth?: IAccountVerificationSessionPreAuth | null;
+  three_ds?: IAccountVerificationSessionThreeDS | null;
+  issuer?: IAccountVerificationSessionIssuer | null;
+  created_at: string;
+  updated_at: string; 
+};
+
+export interface IAccountCreateOpts {
+  holder_id: string;
+  metadata?: {};
+};
+
+export interface IACHCreateOpts extends IAccountCreateOpts {
+  ach: IAccountACH;
+};
+
+export interface ILiabilityCreateOpts extends IAccountCreateOpts {
+  liability: {
+    mch_id: string;
+    account_number?: string;
+    number?: string;
+  }
+};
+
+export interface IAccountListOpts<T extends TAccountExpandableFields> extends IResourceListOpts {
+  status?: string | null;
+  type?: string | null;
+  holder_id?: string | null;
+  expand?: T[];
+  'liability.mch_id'?: string | null;
+  'liability.type'?: TAccountLiabilityTypes | null;
+  'liability.ownership'?: TAccountOwnership | null;
+};
+
+export interface IAccountWithdrawConsentOpts {
+  type: 'withdraw';
+  reason: 'holder_withdrew_consent' | null;
+};
+
+export interface IAccount {
+  id: string;
+  holder_id: string;
+  status: TAccountStatuses;
+  type: TAccountTypes | null;
+  ach?: IAccountACH | null;
+  liability?: IAccountLiability | null;
+  products: TAccountProducts[];
+  restricted_products: TAccountProducts[];
+  subscriptions?: TAccountSubscriptionTypes[];
+  available_subscriptions?: TAccountSubscriptionTypes[];
+  restricted_subscriptions?: TAccountSubscriptionTypes[];
+  sensitive?: string | IAccountSensitive | null;
+  balance?: string | IAccountBalance | null;
+  card_brand?: string | IAccountCardBrand | null;
+  payoff?: string | IAccountPayoff | null;
+  transactions?: string | IAccountTransaction[] | null;
+  update?: string | IAccountUpdate | null;
+  latest_verification_session?: string | IAccountVerificationSession | null;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+  metadata: {} | null;
+};
+
+export * from './externalTypes'
