+++ src/resources/Entity/types.ts
@@ -1,3 +1,9 @@
+import type {
+  IResourceError,
+  TResourceStatus,
+  IResourceListOpts,
+} from "../../resource";
+
 export const CreditReportBureaus = {
   experian: 'experian',
   equifax: 'equifax',
@@ -104,4 +110,283 @@ export interface IEntityVerificationPhone {
 export interface IEntityVerification {
   identity?: IEntityVerificationIdentity;
   phone?: IEntityVerificationPhone;
+};
+
+export interface IEntityConnect {
+  id: string;
+  entity_id: string;
+  status: TResourceStatus;
+  accounts: string[] | null;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export const CreditScoresModel = {
+  vantage_4: 'vantage_4',
+  vantage_3: 'vantage_3',
+};
+
+export type TCreditScoresModel = keyof typeof CreditScoresModel;
+
+export interface IEntityCreditScoresFactorsType {
+  code: string,
+  description: string,
+};
+
+export interface IEntityCreditScoresType {
+  score: number,
+  source: TCreditReportBureaus,
+  model: TCreditScoresModel,
+  factors: IEntityCreditScoresFactorsType[],
+  created_at: string,
+};
+
+export interface IEntityCreditScores {
+  id: string,
+  entity_id: string,
+  status: TResourceStatus,
+  scores: IEntityCreditScoresType[] | null,
+  error: IResourceError | null,
+  created_at: string,
+  updated_at: string,
+};
+
+export interface IEntityIdentity {
+  id: string;
+  entity_id: string;
+  status: TResourceStatus;
+  identities: IEntityIdentityType[];
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export const EntityProductTypeStatuses = {
+  unavailable: 'unavailable',
+  available: 'available',
+  restricted: 'restricted',
+};
+
+export type TEntityProductTypeStatuses = keyof typeof EntityProductTypeStatuses;
+
+export const EntityProductType = {
+  connect: 'connect',
+  credit_score: 'credit_score',
+  identity: 'identity',
+};
+
+export type TEntityProductType = keyof typeof EntityProductType;
+
+export interface IEntityProduct {
+  id: string;
+  name: string;
+  status: TEntityProductTypeStatuses;
+  status_error: IResourceError | null;
+  latest_request_id: string | null;
+  is_subscribable: boolean;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IEntityProductListResponse {
+  connect?: IEntityProduct;
+  credit_score?: IEntityProduct;
+  identity?: IEntityProduct;
+};
+
+export const EntitySubscriptionNames = {
+  connect: 'connect',
+  credit_score: 'credit_score',
+};
+
+export type TEntitySubscriptionNames = keyof typeof EntitySubscriptionNames;
+
+export const EntitySubscriptionStatuses = {
+  active: 'active',
+  inactive: 'inactive',
+};
+
+export type TEntitySubscriptionStatuses = keyof typeof EntitySubscriptionStatuses;
+
+export interface IEntitySubscription {
+  id: string;
+  name: TEntitySubscriptionNames;
+  status: TEntitySubscriptionStatuses;
+  latest_request_id: string | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export interface IEntitySubscriptionResponse {
+  connect?: IEntitySubscription;
+  credit_score?: IEntitySubscription;
+};
+
+export interface IEntitySubscriptionCreateOpts {
+  enroll: TEntitySubscriptionNames;
+};
+
+export const EntityVerificationSessionStatuses = {
+  pending: 'pending',
+  in_progress: 'in_progress',
+  verified: 'verified',
+  failed: 'failed',
+};
+
+export type IEntityVerificationSessionStatuses = keyof typeof EntityVerificationSessionStatuses;
+
+export const EntityVerificationSessionMethods = {
+  sms: 'sms',
+  sna: 'sna',
+  byo_sms: 'byo_sms',
+  byo_kyc: 'byo_kyc',
+  kba: 'kba',
+  element: 'element',
+  method_verified: 'method_verified',
+};
+
+export type IEntityVerificationSessionMethods = keyof typeof EntityVerificationSessionMethods;
+
+export const EntityVerificationSessionTypes = {
+  phone: 'phone',
+  identity: 'identity',
+};
+
+export type IEntityVerificationSessionTypes = keyof typeof EntityVerificationSessionTypes;
+
+export interface IEntitySmsVerification {
+  timestamp: string;
+};
+
+export interface IEntitySmsVerificationUpdate {
+  sms_code: string;
+};
+
+export interface IEntitySnaVerification {
+  urls: string;
+};
+
+export interface IEntityByoKycVerification {
+  authenticated: boolean;
+};
+
+export interface IEntityKbaVerificationAnswer {
+  id: string;
+  text: string;
+};
+
+export interface IEntityKbaVerificationAnswerUpdate {
+  question_id: string;
+  answer_id: string;
+};
+
+export interface IEntityKbaVerificationQuestion {
+  selected_answer?: string;
+  id: string;
+  text: string;
+  answers: IEntityKbaVerificationAnswer[];
+};
+
+export interface IEntityKbaVerification {
+  questions: IEntityKbaVerificationQuestion[];
+  authenticated: boolean;
+};
+
+export interface IEntityVerificationSessionCreateOpts {
+  type: IEntityVerificationSessionTypes;
+  method: IEntityVerificationSessionMethods;
+  sms?: {};
+  sna?: {};
+  byo_sms?: IEntitySmsVerification;
+  byo_kyc?: {};
+  kba?: {};
+};
+
+export interface IEntityVerificationSessionUpdateOpts {
+  type: IEntityVerificationSessionTypes;
+  method: IEntityVerificationSessionMethods;
+  sms?: IEntitySmsVerificationUpdate;
+  sna?: {};
+  kba?: IEntityKbaVerificationAnswerUpdate;
+};
+
+export interface IEntityVerificationSession {
+  id: string;
+  entity_id: string;
+  status: IEntityVerificationSessionStatuses;
+  type: IEntityVerificationSessionTypes;
+  method: IEntityVerificationSessionMethods;
+  sms?: IEntitySmsVerification | null;
+  sna?: IEntitySnaVerification | null;
+  byo_sms?: IEntitySmsVerification | null;
+  byo_kyc?: IEntityByoKycVerification | null
+  kba?: IEntityKbaVerification | null;
+  element?: IEntityKbaVerification | null;
+  error: IResourceError | null;
+  created_at: string;
+  updated_at: string;
+};
+
+export const EntityExpandableFields = {
+  connect: 'connect',
+  credit_score: 'credit_score',
+  identity_latest_verification_session: 'identity.latest_verification_session',
+  phone_latest_verification_session: 'phone.latest_verification_session',
+} as const;
+
+export type TEntityExpandableFields = typeof EntityExpandableFields[keyof typeof EntityExpandableFields];
+
+export interface IEntityCreateOpts {
+  type: TEntityTypes;
+  address?: IEntityAddress | null;
+  metadata?: {} | null;
+};
+
+export interface IIndividualCreateOpts extends IEntityCreateOpts {
+  type: 'individual';
+  individual: Partial<IEntityIndividual>;
+};
+
+export interface ICorporationCreateOpts extends IEntityCreateOpts {
+  type: 'corporation';
+  corporation: Partial<IEntityCorporation>;
+};
+
+export interface IEntityUpdateOpts {
+  address?: IEntityAddress;
+  corporation?: Partial<IEntityCorporation>;
+  individual?: Partial<IEntityIndividual>;
+};
+
+export interface IEntityListOpts<T extends TEntityExpandableFields> extends IResourceListOpts {
+  status?: string | null;
+  type?: string | null;
+  expand?: T[];
+};
+
+export interface IEntityWithdrawConsentOpts {
+  type: 'withdraw',
+  reason: 'entity_withdrew_consent' | null,
+};
+
+export interface IEntity {
+  id: string;
+  type: TEntityTypes | null;
+  individual?: IEntityIndividual | null;
+  corporation?: IEntityCorporation | null;
+  address: IEntityAddress | {};
+  status: TEntityStatuses;
+  error: IResourceError | null;
+  metadata: {} | null;
+  products?: TEntityProductType[];
+  restricted_products?: TEntityProductType[];
+  subscriptions?: TEntitySubscriptionNames[];
+  available_subscriptions?: TEntitySubscriptionNames[];
+  restricted_subscriptions?: TEntitySubscriptionNames[];
+  verification?: IEntityVerification | null;
+  connect?: string | IEntityConnect | null;
+  credit_score?: string| IEntityCreditScores | null;
+  created_at: string;
+  updated_at: string;
 };
